Completed the bidding app
Fixed the bugs from backened on null data empty fields etx. Designed the UI, Added Admin CRUD for AuctionController and ProductsController. Did soft delete on both products table and auctions table. Added the AIZ notifications and made the UI dynamic. On product delete also deleted the auctions that was related to the product. On auction delete deleted the biddings that were realted to it. Added paginate.

Co-Authored-By: Mahmudur Rahman <hridoymahmud71@gmail.com>
Co-Authored-By: abdullah1404111 <32268315+abdullah1404111@users.noreply.github.com>